Allow runtest.{rc,sh} to handle subdirectories.
This allows tests consisting of multiple files, including .glue.c files,
so we can check calling convention compliance.

+++ test/runtest.sh
@@ -1,15 +1,25 @@
 #!/bin/sh
-export PATH=.:$PATH
-export MYR_MC=../6/6m
-export MYR_MUSE=../muse/muse
+export PATH=$(pwd):$PATH
+export MYR_MC=$(cd ..; pwd)/6/6m
+export MYR_MUSE=$(cd ..; pwd)/muse/muse
 ARGS=$*
 NFAILURES=0
 NPASSES=0
 
 build() {
-	rm -f out/$1 out/$1.o out/$1.s out/$1.use
-	mkdir -p out
-	../obj/mbld/mbld -Bnone -o 'out' -b $1 -I../obj/lib/std -I../obj/lib/sys -I../obj/lib/regex -r../rt/_myrrt.o $1.myr
+	dir=$(echo $1 | egrep -o '.*/')
+	if [ -z $dir ]; then
+		rm -f out/$1 out/$1.o out/$1.s out/$1.use
+		mkdir -p out
+		../obj/mbld/mbld -Bnone -o 'out' -b $1 -I../obj/lib/std -I../obj/lib/sys -I../obj/lib/regex -r../rt/_myrrt.o $1.myr
+	else
+		target=$(echo $1 | egrep -o '[^/]*$')
+		top=$(pwd)
+		mkdir -p out/$dir
+		cd $dir
+		$top/../obj/mbld/mbld -Bnone -o $top/out/$dir -I$top/../obj/lib/std -I$top/../obj/lib/sys -I$top/../obj/lib/regex -r$top/../rt/_myrrt.o clean
+		$top/../obj/mbld/mbld -Bnone -o $top/out/$dir -I$top/../obj/lib/std -I$top/../obj/lib/sys -I$top/../obj/lib/regex -r$top/../rt/_myrrt.o :$target
+	fi
 }
 
 pass() {
@@ -95,7 +105,9 @@ B() {
 		args="$1"; shift
 	fi
 	echo "test $test <<{!"
+	here=$(pwd)
 	build $test
+	cd $here
 	case $type in
 	"E")  expectstatus "$test" "$res";;
 	"P")  expectprint "$test" "$res";;
@@ -111,12 +123,14 @@ F() {
 	fi
 
 	echo "test $1 <<{!"
-	(build $1) > /dev/null 2>1
+	here=$(pwd)
+	(build $1) > /dev/null 2>&1
 	if [ $? -eq '1' ]; then
 		pass $1
 	else
 		fail $1
 	fi
+	cd $here
 }
 
 echo "MTEST $(egrep '^[BF]' tests | wc -l)"
